+++ src/pc/video_vga.h
@@ -6,3 +6,9 @@
     IBM VGA (1987) emulation
     Incl. Mode 4,5,6,7,8,9,10,11,12,13h...X,Y...Blah blah!
 */
+
+// Enumerate VGA registers
+typedef enum vga_registers
+{
+
+};